move CopyUrlForm to src/pages/settings/formgroup
